Analysis vizualization tables now use Tree Mode
