--- client/src/containers/analysis-visualization/analysis-table/single-indicator-table/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './component';
